requirements.txt
configparser==3.7.3
Django==2.1.7
djangorestframework==3.9.2
pycrypto==2.6.1
pytz==2018.9
